[WIP] Refactoring the application
This commit introduces a significant refactor of the overall
application.
We now rely directly on a Runner, and a lot of boilerplate has been
removed. Additionally, the application is only taking into account
method calls after the PRE_INITIAL phase, which therefore reduces a lot
the execution time.
The javacc dependency has been dropped, by removing the SettingsReader
part of the application in JBSE.

Still a WIP:
- We need to cleanup our application code significantly
- We can remove the "apps" part of JBSE, since it is no longer needed.

In any case, this is a good starting (and working!) point.

Signed-off-by: Alessandro Pellegrini <pellegrini@diag.uniroma1.it>