Consider all test programs in grouping
This requires to use the dump of all test programs from method
enumeration phase during the analysis.
Code could be refactored to recompute the set, but this would put a
non-necessary burden on the grouping stage, and we don't really know
whether these phases will be retained in the future.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>